tight-binding-package/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mypkg"
version = "0.1.0"
description = "Tight-binding / NEGF research code"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.3",
  "pydantic>=2.12",
  "pyyaml>=6.0",
  "pandas>=2.3",
  "scipy>=1.15",
  "numba>=0.63",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "ruff>=0.1",
  "mypy>=1.7",
  "ipython>=8",
  "ipykernel>=6",
  "matplotlib>=3.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
